requirements.txt
appdirs>=1.4.0
numpy
gunicorn
django-heroku
pytz
pandas
